feat: display goods by good group, refactor the page related components into separate files
